requirements.txt
tweepy==4.14.0
openai==1.52.0
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.1.0
web3==6.11.3
requests==2.31.0
google-generativeai
